requirements.txt
PyGithub==1.34